Use uint16_t for MCC and MNC
As Dieter correctly points out, MCC an MNC can very well be lager than
255 ;)

+++ openbsc/src/osmo-ganc/ganc_data.h
@@ -47,8 +47,8 @@ struct gan_peer {
 };
 
 struct ganc_net {
-	uint8_t country_code;
-	uint8_t network_code;
+	uint16_t country_code;
+	uint16_t network_code;
 
 	uint16_t timer[_NUM_GANC_TIMER];
